LayoutTests/fast/dom/shadow/selection-in-shadow.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/polyfill.js"></script>
</head>
<body>

<div id="container"></div>

<script>
function moveTo(element) {
    var x = element.offsetLeft + element.offsetWidth / 2;
    var y = element.offsetTop + element.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
}

function dragFromTo(elementFrom, elementTo) {
    moveTo(elementFrom);
    eventSender.mouseDown();
    moveTo(elementTo);
    eventSender.mouseUp();
}

var container = document.getElementById('container');

var shadowRoot1 = new WebKitShadowRoot(container);
shadowRoot1.innerHTML =
    '<div id="div1" title="div1">DIV1</div>' +
    '<div id="div2" title="div2">DIV2</div>' +
    '<div id="div3" title="div3">DIV3</div>';
var div1 = shadowRoot1.getElementById('div1');
var div2 = shadowRoot1.getElementById('div2');
var div3 = shadowRoot1.getElementById('div3');

var shadowRoot2 = new WebKitShadowRoot(div2);
shadowRoot2.innerHTML =
    '<div id="div4" title="div4">DIV4</div>' +
    '<div id="div5" title="div5">DIV5</div>' +
    '<div id="div6" title="div6">DIV6</div>';
var div4 = shadowRoot2.getElementById('div4');
var div5 = shadowRoot2.getElementById('div5');
var div6 = shadowRoot2.getElementById('div6');

var shadowRoot3 = new WebKitShadowRoot(div5);
shadowRoot3.innerHTML =
    '<div id="div7" title="div7">DIV7</div>' +
    '<div id="div8" title="div8">DIV8</div>' +
    '<div id="div9" title="div9">DIV9</div>';
var div7 = shadowRoot3.getElementById('div7');
var div8 = shadowRoot3.getElementById('div8');
var div9 = shadowRoot3.getElementById('div9');


dragFromTo(div4, div6);
var selectionFromWindow = window.getSelection();
var selectionFromShadowRoot1 = shadowRoot1.selection;
var selectionFromShadowRoot2 = shadowRoot2.selection;
var selectionFromShadowRoot3 = shadowRoot3.selection;

// Since there are these bugs, shadowRoot.selection won't return the correct selection now.
// Currently it is the same as window.getSelection().
// FIXME: VisibleSelection in ShadowRoot is not intuitive.
// http://wkb.ug/82683
// FIXME: ShadowRoot.selection should return selection whose range is in a shadow tree.
// http://wkb.ug/82698

shouldBe('selectionFromWindow.anchorNode', 'selectionFromShadowRoot1.anchorNode');
shouldBe('selectionFromWindow.anchorNode', 'selectionFromShadowRoot2.anchorNode');
shouldBe('selectionFromWindow.anchorNode', 'selectionFromShadowRoot3.anchorNode');

shouldBe('selectionFromWindow.focusNode', 'selectionFromShadowRoot1.focusNode');
shouldBe('selectionFromWindow.focusNode', 'selectionFromShadowRoot2.focusNode');
shouldBe('selectionFromWindow.focusNode', 'selectionFromShadowRoot3.focusNode');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
