LayoutTests/platform/chromium/fast/dom/shadow/selection-in-shadow-expected.txt

PASS selectionFromWindow.anchorNode is selectionFromShadowRoot1.anchorNode
PASS selectionFromWindow.anchorNode is selectionFromShadowRoot2.anchorNode
PASS selectionFromWindow.anchorNode is selectionFromShadowRoot3.anchorNode
PASS selectionFromWindow.focusNode is selectionFromShadowRoot1.focusNode
PASS selectionFromWindow.focusNode is selectionFromShadowRoot2.focusNode
PASS selectionFromWindow.focusNode is selectionFromShadowRoot3.focusNode
PASS successfullyParsed is true

TEST COMPLETE
